requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic[email]==2.4.2
pillow==10.1.0
numpy==1.26.2
boto3==1.33.2

# Для продакшена
gunicorn==21.2.0
alembic==1.12.1
prometheus-client==0.17.1
sentry-sdk==1.32.0
tenacity==8.2.3
bcrypt==4.0.1
passlib==1.7.4
httpx==0.25.0 